Add forkexec system call
forkexec is a combination of fork and exec where the arguments are same
as that of exec. This syscall creates a new process and loads the
executable from the path and schedule it to run.

forkexec loads the process to the vaddr specified in the elf program
header. Make sure all programs are built with different base address.